// one-cgiar-front/src/app/pages/stages-container/stages/shared/components/general-information/interfaces/general-info-patch-body.interface.ts
export interface GeneralInfoPatchBody {
    name: string;
    action_area_description: string;
    action_area_id: number;
    generalInformationId: number;
    table_name: string;
    col_name: string;
    active: boolean;
    budgetId: number;
    is_global: number;
    budget_value: string;
    acronym:string,
  }